renamed Lycan to Werewolf and vice-versa
