+++ Games/types/Mafia/const/Priority.js
@@ -19,8 +19,8 @@ const PRIORITY_BLOCK_VISITORS = -96;
 const PRIORITY_NIGHT_ROLE_BLOCKER = -95;
 const PRIORITY_NIGHT_NURSE = -95;
 
-const PRIORITY_CLEANSE_WEREWOLF_VISITORS = -92;
-const PRIORITY_KILL_LYCAN_VISITORS_ENQUEUE = -92;
+const PRIORITY_CLEANSE_LYCAN_VISITORS = -92;
+const PRIORITY_KILL_WEREWOLF_VISITORS_ENQUEUE = -92;
 const PRIORITY_VISITORS_ENQUEUE = -92;
 
 const PRIORITY_NIGHT_SAVER = -90;
@@ -92,8 +92,8 @@ module.exports = {
   PRIORITY_NIGHT_ROLE_BLOCKER,
   PRIORITY_NIGHT_NURSE,
   PRIORITY_BLOCK_VISITORS,
-  PRIORITY_CLEANSE_WEREWOLF_VISITORS,
-  PRIORITY_KILL_LYCAN_VISITORS_ENQUEUE,
+  PRIORITY_CLEANSE_LYCAN_VISITORS,
+  PRIORITY_KILL_WEREWOLF_VISITORS_ENQUEUE,
   PRIORITY_VISITORS_ENQUEUE,
   PRIORITY_NIGHT_SAVER,
 
@@ -125,7 +125,7 @@ module.exports = {
   PRIORITY_MAFIA_KILL,
 
   PRIORITY_KILL_DEFAULT,
-  PRIORITY_KILL_LYCAN_VISITORS,
+  PRIORITY_KILL_WEREWOLF_VISITORS,
 
   PRIORITY_NIGHT_REVIVER,
