+++ Games/types/Mafia/roles/Cult/Werewolf.js
@@ -1,0 +1,17 @@
+const Role = require("../../Role");
+
+module.exports = class Werewolf extends Role {
+  constructor(player, data) {
+    super("Werewolf", player, data);
+
+    this.alignment = "Cult";
+    this.cards = [
+      "VillageCore",
+      "WinWithCult",
+      "MeetingCult",
+      "BitingWolf",
+      "FullMoonInvincible",
+      "CauseFullMoons",
+    ];
+  }
+};
