+++ Games/types/Mafia/roles/Village/Priest.js
@@ -8,8 +8,8 @@ module.exports = class Priest extends Role {
     this.cards = [
       "VillageCore",
       "WinWithVillage",
-      "CleanseWerewolfVisitors",
-      "KillLycanVisitors",
+      "CleanseLycanVisitors",
+      "KillWerewolfVisitors",
     ];
     this.immunity.wolfBite = 1;
   }
